fix SLF4J startup
Underlying libraries, such as the Java Riak driver, rely on SLF4J, a
logging library. Previously, upon startup, the initialization of SLF4J
would fail if SLF4J was not installed and configured properly. This had
the side effect of potentially disrupting other loggers, such as Jetty's
default logging capabilities.

Here we introduce a new dependency, SLF4J and provide a default
configuration for it.

Fixes: #7